requirements.txt
pytest>=8.2.2
selenium>=4.22.0
python-dotenv>=1.0.1
webdriver-manager>=4.0.2
behave>=1.3.1
pytest_bdd>=7.1.2
pytest-html>=4.1.1
requests>=2.32.3
Faker>=25.0.0
jsonschema>=4.22.0